+++ src/protocol/config.c
@@ -128,20 +128,20 @@ void _z_str_intmap_onto_str(char *dst, size_t dst_len, const _z_str_intmap_t *s,
     for (size_t i = 0; i < argc; i++) {
         char *v = _z_str_intmap_get(s, argv[i]._key);
         if (v != NULL) {
-            if (len > 0) {
+            if (len > (size_t)0) {
                 (void)strncat(dst, &lsep, 1);  // List separator
-                len = len - 1;
+                len = len - (size_t)1;
             }
 
-            if (len > 0) {
+            if (len > (size_t)0) {
                 (void)strncat(dst, argv[i]._str, len);  // Key
                 len = len - strlen(argv[i]._str);
             }
-            if (len > 0) {
+            if (len > (size_t)0) {
                 (void)strncat(dst, &ksep, 1);  // KeyValue separator
-                len = len - 1;
+                len = len - (size_t)1;
             }
-            if (len > 0) {
+            if (len > (size_t)0) {
                 (void)strncat(dst, v, len);  // Value
                 len = len - strlen(v);
             }
